+++ CHANGES.md
@@ -1,3 +1,9 @@
+## Feb 13, 2012
+
+* Version 0.4
+* Propagate `connection.remoteAddress` to scripts and patches. Scripts can access this via
+  `{remoteAddress}` and patches via `params.remoteAddress`.
+
 ## Feb 10, 2012
 
 * Set `application/json` on JSON responses.